all main logic is READY!!!

+++ app/src/main/java/com/example/findme/presentation/account/AccountVM.kt
@@ -71,16 +71,19 @@ class AccountVM @Inject constructor(
     private var _editState = MutableLiveData<String?>()
     val editState: LiveData<String?> get() = _editState
 
-    fun edit(lifecycleOwner: LifecycleOwner, login: String, name: String, surname: String, password: String, avatar: Uri?){
+    fun edit(lifecycleOwner: LifecycleOwner, login: String, name: String, surname: String, password: String, avatar: String?){
         waitForError(lifecycleOwner)
-        if(avatar != null) {
-            storageUC.changeAvatar(avatar, login)
+        if(avatar != "null" && avatar != null && !avatar.startsWith("https://", ignoreCase = true)) {
+            storageUC.changeAvatar(avatar.toUri(), login)
             storageUC.storageState.observe(lifecycleOwner){ avatarUrl ->
                 if(avatarUrl != null){
                     editAccDB(lifecycleOwner, login, name, surname, password, avatarUrl)
                 }
             }
-        }else{
+        }else if(avatar!!.startsWith("https://", ignoreCase = true)){
+            editAccDB(lifecycleOwner, login, name, surname, password, avatar)
+        }
+        else{
             storageUC.deleteAvatar(login)
             storageUC.storageState.observe(lifecycleOwner){ result ->
                 editAccDB(lifecycleOwner, login, name, surname, password, "")
@@ -100,6 +103,7 @@ class AccountVM @Inject constructor(
     }
 
     private fun waitForError(lifecycleOwner: LifecycleOwner) {
+        _error.value = ErrorStates.NULL
         storageUC.errorStorage.observe(lifecycleOwner){ error ->
             _error.value = error!!
         }
